Re-use any previous matching transcode instead of re-running
This means we'll use any previous transcode of the same file.  This
should be fine unless for some reason we really need a new transcode
(e.g., new ffmpeg version) but then probably we should fix the old one
instead.  It also means that the metadata won't match, but maybe that's
okay too.  It will certainly safe disk space and cpu time.