Change pgDBPort default configuration
Currently, the hardcoded socket is used if no host, port, or socket are
specified in the config file. I think it is unlikely that this hardcoded socket
would be used without specification. (I ran into this problem myself.)
Therefore, I'm changing the way the host and port defaults are determined.